Gestures handles exceptions properly
Previously whenever event handlers threw any exception, it broke the
internal
state of TouchScript, which then started printing many error messages.
Now exceptions are handled and printed as a log message to Unity
console.

This could be optional in future, because now it prevents developer from
actually debugging the code where the exceptions are thrown.