Add optional non-simultaneous transform restriction
Can now disallow multiple types of transformations to occur each time a
gesture is used.

For example, when a gesture is used to rotate it cannot scale or
translate. Once the gesture ends it is reset and can again be used for
any type of transformation.

There is a checkbox in the inspector to select this option. It is under
Advanced > General > Simultaneous Transforms.

It works for both One- and Two- Point transformations, including
two-finger translation, 2+ finger transformations, and clustered
transforms.